refactor(domain/lotto): simplify Lotto class and improve validation clarity
